feature/Added the utility to print colored text (#55)
## Description
I added the utility to print colored text. It is a function called
ColorText in the utils package where you can pass a string and a color
from the stdOutColor struct, and it will return a formatted string that
adds the color and you can print it in the terminal.